add put, delete and post for employees and managers

+++ WebApiAssignment/Controllers/EmployeesController.cs
@@ -19,37 +19,38 @@ namespace WebApiAssignment.Controllers
         {
             _context = context;
 
-            _context.Employees.RemoveRange(_context.Employees);
-            _context.SaveChanges();
+            if (Startup.ResetEmployeeDatabase == true)
+            {
+                Startup.ResetEmployeeDatabase = false;
+                _context.Employees.RemoveRange(_context.Employees);
+                _context.SaveChanges();
 
-            _context.Database.ExecuteSqlCommand("UPDATE sqlite_sequence SET seq = 1 WHERE name = 'employee';");
-            _context.SaveChanges();
-
-            _context.Employees.Add(new Employee
-            {
-                Name = "Vighnesh",
-                Age = 21,
-                Salary = 10000000
-            });
-            _context.Employees.Add(new Employee
-            {
-                Name = "Shubham",
-                Age = 30,
-                Salary = 1232123
-            });
-            _context.Employees.Add(new Employee
-            {
-                Name = "Omkar",
-                Age = 32,
-                Salary = 131232
-            });
-            _context.Employees.Add(new Employee
-            {
-                Name = "Bhanu",
-                Age = 45,
-                Salary = 12313200
-            });
-            _context.SaveChanges();
+                _context.Employees.Add(new Employee
+                {
+                    Name = "Vighnesh",
+                    Age = 21,
+                    Salary = 10000000
+                });
+                _context.Employees.Add(new Employee
+                {
+                    Name = "Shubham",
+                    Age = 30,
+                    Salary = 1232123
+                });
+                _context.Employees.Add(new Employee
+                {
+                    Name = "Omkar",
+                    Age = 32,
+                    Salary = 131232
+                });
+                _context.Employees.Add(new Employee
+                {
+                    Name = "Bhanu",
+                    Age = 45,
+                    Salary = 12313200
+                });
+                _context.SaveChanges();
+            }
         }
 
         [HttpGet]
@@ -58,17 +59,56 @@ namespace WebApiAssignment.Controllers
             return await _context.Employees.ToListAsync();
         }
 
-        [HttpGet("{id}")]
-        public async Task<ActionResult<Employee>> GetEmployee(long id)
-        {
-            var employees = await _context.Employees.FindAsync(id);
 
-            if (employees == null)
+        [HttpPost]
+        public async Task<ActionResult> AddEmployee([FromBody] Employee employee)
+        {
+            await _context.Employees.AddAsync(employee);
+            await _context.SaveChangesAsync();
+
+            return Created("Employee added to DataBase", employee);
+        }
+
+        [HttpGet("{id}")]
+        public async Task<ActionResult<Employee>> GetEmployee(int id)
+        {
+            var employee = await _context.Employees.FindAsync(id);
+
+            if (employee == null)
             {
                 return NotFound();
             }
 
-            return employees;
+            return employee;
+        }
+
+        [HttpPut("{id}")]
+        public async Task<ActionResult> UpdateEmployee(int id, [FromBody] Employee employee)
+        {
+            if (await _context.Employees.FindAsync(id) is null)
+                return BadRequest();
+
+            if (id != employee.Id)
+                return BadRequest();
+
+            _context.Entry(employee).State = EntityState.Modified;
+            await _context.SaveChangesAsync();
+
+            return NoContent();
+        }
+
+        [HttpDelete("{id}")]
+        public async Task<ActionResult> DeleteEmployee(int id)
+        {
+            var employee = await _context.Employees.FindAsync(id);
+
+            if (employee is null)
+                return BadRequest();
+
+            _context.Employees.Remove(employee);
+            await _context.SaveChangesAsync();
+
+            return Ok("Employee removed sucessfully!");
         }
     }
 }
